requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pytest
redis
aiogram